Isolate the RegisterUser and Password classes
* Move the RegisterUser and Password classes to a separate module for
  ease of testing and maintenance.
