+++ test/requests/test_login_local.py
@@ -1,5 +1,5 @@
 import requests
-from wqflask import user_manager
+from wqflask.user_manager import RegisterUser
 from parameterized import parameterized
 from parametrized_test import ParametrizedTest
 
@@ -19,8 +19,7 @@ class TestLoginLocal(ParametrizedTest):
             "password": "test_password",
             "password_confirm": "test_password"
         }
-        user_manager.basic_info = lambda : { "basic_info": "basic" }
-        user_manager.RegisterUser(data)
+        RegisterUser(data)
 
     
     @parameterized.expand([
